Move BaseHandler to separate module
This allows other modules to import it without circular
dependencies.
