+++ AdminServer/appscale/admin/__init__.py
@@ -29,6 +29,7 @@ from tornado.escape import json_encode
 from tornado.ioloop import IOLoop
 from . import utils
 from . import constants
+from .base_handler import BaseHandler
 from .constants import (
   CustomHTTPError,
   OperationTimeout,
@@ -136,38 +137,6 @@ def wait_for_delete(operation_id, http_port):
     yield gen.sleep(1)
 
   operation.finish()
-
-
-class BaseHandler(web.RequestHandler):
-  """ A handler with helper functions that other handlers can extend. """
-  def authenticate(self):
-    """ Ensures requests are authenticated.
-
-    Raises:
-      CustomHTTPError if the secret is invalid.
-    """
-    if 'AppScale-Secret' not in self.request.headers:
-      message = 'A required header is missing: AppScale-Secret'
-      raise CustomHTTPError(HTTPCodes.UNAUTHORIZED, message=message)
-
-    if self.request.headers['AppScale-Secret'] != options.secret:
-      raise CustomHTTPError(HTTPCodes.UNAUTHORIZED, message='Invalid secret')
-
-  def write_error(self, status_code, **kwargs):
-    """ Writes a custom JSON-based error message.
-
-    Args:
-      status_code: An integer specifying the HTTP error code.
-    """
-    details = {'code': status_code}
-    if 'exc_info' in kwargs:
-      error = kwargs['exc_info'][1]
-      try:
-        details.update(error.kwargs)
-      except AttributeError:
-        pass
-
-    self.finish(json_encode({'error': details}))
 
 
 class VersionsHandler(BaseHandler):
